Rename and move some Node classes
