+++ integrations/webuntis/src/main/kotlin/me/snoty/integration/untis/WebUntisIntegration.kt
@@ -1,12 +1,13 @@
 package me.snoty.integration.untis
 
 import me.snoty.integration.common.annotation.RegisterNode
-import me.snoty.integration.common.fetch.AbstractIntegrationFetcher
 import me.snoty.integration.common.fetch.FetchContext
+import me.snoty.integration.common.fetch.fetchContext
 import me.snoty.integration.common.model.NodePosition
 import me.snoty.integration.common.wiring.*
 import me.snoty.integration.common.wiring.data.EmitNodeOutputContext
 import me.snoty.integration.common.wiring.data.IntermediateData
+import me.snoty.integration.common.wiring.node.NodeHandler
 import me.snoty.integration.untis.model.UntisExam
 import me.snoty.integration.untis.request.getExams
 import org.jobrunr.jobs.context.JobContext
@@ -18,10 +19,10 @@ import org.slf4j.Logger
 	position = NodePosition.START,
 	settingsType = WebUntisSettings::class
 )
-open class WebUntisFetcher(
+class WebUntisIntegration(
 	override val nodeHandlerContext: NodeHandlerContext,
 	private val untisAPI: WebUntisAPI = WebUntisAPIImpl(nodeHandlerContext.httpClient())
-) : AbstractIntegrationFetcher() {
+) : NodeHandler {
 	override val settingsClass = WebUntisSettings::class
 
 	context(FetchContext, NodeHandlerContext)
@@ -47,7 +48,7 @@ open class WebUntisFetcher(
 	context(NodeHandlerContext, EmitNodeOutputContext)
 	override suspend fun process(logger: Logger, node: Node, input: IntermediateData) {
 		val jobContext: JobContext = input.get()
-		val fetchContext = progress(logger, jobContext, 1)
+		val fetchContext = fetchContext(logger, jobContext, 1)
 
 		iterableStructOutput(fetchContext) {
 			fetchExams(node, logger)
