+++ integrations/webuntis/src/main/kotlin/me/snoty/integration/untis/WebUntisSettings.kt
